+++ docs/changelog.rst
@@ -19,6 +19,7 @@ Changelog
 - Added specific check for declared :attr:`~import_export.options.ResourceOptions.import_id_fields` not in dataset
   (#1735)
 - Updated Admin integration documentation to clarify how to save custom form values (#1746)
+- Fixed handling of :attr:`~import_export.exceptions.FieldError` during Admin import (#17)
 
 4.0.0-beta.2 (2023-12-09)
 -------------------------